Refactor to move from case detail into a component
Jira ticket: CAMS-338
